Refactor ue_movement endpoint url
