+++ backend/app/app/api/api_v1/endpoints/__init__.py
@@ -1,7 +1,7 @@
 from .login import router
 from .users import router
 from .utils import router
-from .ue_movement import router
+from .sim_ue_movement import router
 from .paths import router
 from .gNB import router
 from .Cell import router
