+++ backend/app/app/static/js/dashboard-ues.js
@@ -229,7 +229,7 @@ function api_post_UE_callback(UE_obj, callback) {
 // 
 // 
 function api_get_state_loop_for(UE_supi) {
-    var url = app.api_url + '/ue_movement/state-loop/' + UE_supi;
+    var url = app.api_url + '/sim_ue_movement/state-loop/' + UE_supi;
 
     $.ajax({
         type: 'GET',
